Scoring Logic Moved
The scoreboard no longer references the goals.
Both goals now reference the scoreboard.
IDK if this is more efficient or not....
The scoreboards have many more variables that are being copied than the goals did
But, when doing it this way we no longer update scoreboard every second (regardless of a change in score)