Score and Timer Added to Menu
The scores and timer were added to the menu.

Did not test. The placement is bad (is to be changed.)

Also, Technically not all this would be on a main menu, some would only be found on the pause menu, so this is to be duplicated then catered to the specific menu.